Refactor repository modules and update imports.
