+++ src/gpt_review/repositories/_github.py
@@ -11,7 +11,7 @@ from knack.commands import CommandGroup
 
 from gpt_review._command import GPTCommandGroup
 from gpt_review._review import _summarize_files
-from gpt_review._repository import _RepositoryClient
+from gpt_review.repositories._repository import _RepositoryClient
 
 
 class _GitHubClient(_RepositoryClient):
